feat: Refactor invoice form component and integrate form handling with new structure
