+++ src/app/components/form-component/form-component.component.html
@@ -1,0 +1,128 @@
+<form #newForm [formGroup]="invoiceForm" (ngSubmit)="onSubmit($event)" class="form-container">
+    <section class="navigation-conatiner">
+        <a [routerLink]="['/']" class="back-link rem-md">
+            <img class="back-icon" src="assets/image/icon-arrow-left.svg" alt="back image" />
+            <p> Go back</p>
+        </a>
+        <h1 class="form-title">New Invoice</h1>
+    </section>
+    <section formGroupName="senderAddress" class="group-container">
+        <h2>Bill From</h2>
+        <div class="col-group">
+            <label for="">Street Address</label>
+            <input formControlName="street" type="text" />
+        </div>
+        <div class="row-group">
+            <div class="col-group">
+                <label for="">City</label>
+                <input formControlName="city" type="text" />
+            </div>
+            <div class="col-group">
+                <label for="">Post Code</label>
+                <input formControlName="postCode" type="text" />
+            </div>
+            <div class="col-group">
+                <label for="">Country</label>
+                <input formControlName="country" type="text" />
+            </div>
+        </div>
+    </section>
+
+    <section class="group-container">
+        <h2>Bill To</h2>
+        <div class="col-group">
+            <label for="">Client’s Name</label>
+            <input formControlName="clientName" type="text" />
+        </div>
+        <div class="col-group">
+            <label for="">Client’s Email</label>
+            <input formControlName="clientEmail" type="text" />
+        </div>
+        <div formGroupName="clientAddress" class="col-group">
+            <label for="">Street Address</label>
+            <input formControlName="street" type="text" />
+        </div>
+        <div class="row-group" formGroupName="clientAddress">
+            <div class="col-group">
+                <label for="">City</label>
+                <input formControlName="city" type="text" />
+            </div>
+            <div class="col-group">
+                <label for="">Post Code</label>
+                <input formControlName="postCode" type="text" />
+            </div>
+            <div class="col-group">
+                <label for="">Country</label>
+                <input formControlName="country" type="text" />
+            </div>
+        </div>
+
+        <section class="group-container">
+            <div class="row-group">
+                <div class="col-group">
+                    <label for="">Invoice Date</label>
+
+                    <input formControlName="paymentDue" type="date" />
+                </div>
+                <div class="col-group payment-term-container">
+                    <label for="">Payment Terms</label>
+                    <button (click)="toggleShowTerm()" type="button" class="custom-drop-down">
+                        <span>
+                            {{ getSelectedTermLabel() }}
+                        </span>
+                        <img src="assets/image/icon-arrow-down.svg" />
+                    </button>
+                    <ng-template [ngIf]="showPaymentTerm">
+
+                        <div class="dropdowm-container">
+                            @for (term of paymentTerms; track $index) {
+                            <button (click)="setSelectedTerm(term.id)" type="button" class="custom-drop-down-item"
+                                [ngClass]="{
+                                        'border-b': term.id != 4
+                                        }">
+                                <span>{{term.label}}</span>
+                            </button>
+                            }
+
+                        </div>
+                    </ng-template>
+                </div>
+            </div>
+            <div class="col-group">
+                <label for="">Project Description</label>
+                <input formControlName="description" type="text" />
+            </div>
+        </section>
+
+    </section>
+
+    <section class="group-container">
+        <h2>Item List</h2>
+        <div class="item-list-header">
+            <h3>Item Name</h3>
+            <h3>Qty.</h3>
+            <h3>Price</h3>
+            <h3>Total</h3>
+        </div>
+        <div formArrayName="items">
+            <div *ngFor="let item of items.controls; let i = index" [formGroupName]="i" class="list-item-data">
+                <div class="col-group">
+                    <input formControlName="name" type="text" />
+                </div>
+                <div class="col-group">
+                    <input formControlName="quantity" type="number" type="text" />
+                </div>
+                <div class="col-group">
+                    <input formControlName="price" type="number" type="text" />
+                </div>
+                <div class="col-group">
+                    <input formControlName="total" type="text" [disabled]="true" />
+                </div>
+            </div>
+        </div>
+
+        <button (click)="addItem()" type="button" class="add-invoice-item">
+            + Add New Item
+        </button>
+    </section>
+</form>
