+++ src/app/components/form-component/form-component.component.scss
@@ -1,89 +1,9 @@
 @use './../../../styles/mixins' as *;
 
-
-.new-invoice-wrapper {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    display: flex;
-    flex-direction: row;
-    background-color: rgba(0, 0, 0, 0.5);
-}
-
-app-sidebar {
-    position: absolute;
-    top: 0;
-    left: 0;
-
-}
-
-.wrapper-content {
-    background-color: #fff;
-    width: 616px;
-    height: 100vh;
-    border-top-right-radius: 20px;
-    border-bottom-right-radius: 20px;
-    display: flex;
-    flex-direction: column;
-    padding-left: 103px;
-    overflow: hidden;
-}
-
 .form-container {
     flex: 1;
 }
 
-.form-action {
-    background-color: #fff;
-    padding: 31px 56px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    border-top-right-radius: 20px;
-    border-top: 1px solid rgba(128, 128, 128, 0.11);
-}
-
-.discard-btn {
-    @include light-blue-12;
-    background-color: #F9FAFE;
-    color: #7E88C3;
-    border: none;
-    font-weight: bold;
-    padding: 17px 23px;
-    cursor: pointer;
-    border-radius: 24px;
-}
-
-.save-btn-container {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-}
-
-.save-draft {
-    @include light-blue-12;
-    background-color: #373B53;
-    color: #888EB0;
-    border: none;
-    font-weight: bold;
-    padding: 17px 23px;
-    cursor: pointer;
-    border-radius: 24px;
-}
-
-.save-send {
-    @include light-blue-12;
-    background-color: #7C5DFA;
-    color: #FFFFFF;
-    border: none;
-    font-weight: bold;
-    padding: 17px 23px;
-    cursor: pointer;
-    border-radius: 24px;
-
-}
 
 .form-title {
     padding-top: 56px;
@@ -124,7 +44,7 @@ app-sidebar {
     flex: 1;
 }
 
-.form-container,app-form-component {
+.form-container {
     overflow-y: auto;
 }
 
@@ -228,40 +148,8 @@ label {
     display: none;
 }
 
-@media(max-width: 1024px) {
-    .new-invoice-wrapper {
-        flex-direction: column;
-    }
-
-    app-sidebar {
-        position: fixed;
-        z-index: 10;
-    }
-
-    .wrapper-content {
-        position: absolute;
-        padding-left: 0px;
-        height: 90.4vh;
-        max-width: 85vw;
-        bottom: 0;
-        left: 0;
-    }
-
-}
 
 @media screen and (max-width: 768px) {
-    .wrapper-content {
-        position: absolute;
-        padding-left: 0px;
-        height: 90.4vh;
-        max-width: 100vw;
-        border-radius: 0%;
-    }
-
-    .form-action {
-        border-radius: 0%;
-        padding: 21px 24px;
-    }
 
     .form-container {
         padding: 24px;
